pyproject.toml
[tool.poetry]
name = "fastapi-huella"
version = "1.0.0"
description = "FastAPI application for fingerprint capture and management using ZKTeco devices"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
pyzkfp = "^1.0.0"
Pillow = "^10.1.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:main"
